+++ webrtc/src/webrtc_c/libyuv.cc
@@ -7,6 +7,7 @@
 // libyuv
 #include <libyuv/convert.h>
 #include <libyuv/convert_from.h>
+#include <libyuv/planar_functions.h>
 #include <libyuv/rotate.h>
 #include <libyuv/video_common.h>
 
@@ -85,6 +86,39 @@ WEBRTC_EXPORT int libyuv_I420ToNV12(const uint8_t* src_y,
   return libyuv::I420ToNV12(src_y, src_stride_y, src_u, src_stride_u, src_v,
                             src_stride_v, dst_y, dst_stride_y, dst_uv,
                             dst_stride_uv, width, height);
+}
+
+WEBRTC_EXPORT int libyuv_I420Copy(const uint8_t* src_y,
+                                  int src_stride_y,
+                                  const uint8_t* src_u,
+                                  int src_stride_u,
+                                  const uint8_t* src_v,
+                                  int src_stride_v,
+                                  uint8_t* dst_y,
+                                  int dst_stride_y,
+                                  uint8_t* dst_u,
+                                  int dst_stride_u,
+                                  uint8_t* dst_v,
+                                  int dst_stride_v,
+                                  int width,
+                                  int height) {
+  return libyuv::I420Copy(src_y, src_stride_y, src_u, src_stride_u, src_v,
+                          src_stride_v, dst_y, dst_stride_y, dst_u,
+                          dst_stride_u, dst_v, dst_stride_v, width, height);
+}
+
+WEBRTC_EXPORT int libyuv_NV12Copy(const uint8_t* src_y,
+                                  int src_stride_y,
+                                  const uint8_t* src_uv,
+                                  int src_stride_uv,
+                                  uint8_t* dst_y,
+                                  int dst_stride_y,
+                                  uint8_t* dst_uv,
+                                  int dst_stride_uv,
+                                  int width,
+                                  int height) {
+  return libyuv::NV12Copy(src_y, src_stride_y, src_uv, src_stride_uv, dst_y,
+                          dst_stride_y, dst_uv, dst_stride_uv, width, height);
 }
 
 WEBRTC_EXPORT int libyuv_YUY2ToI420(const uint8_t* src_yuy2,